Move ddsi_discovery_sedp into ddsi_discovery
Trying to move the common parts of SEDP support into a separate file
only confused things, in hindsight.  This moves the contents of that
file back into ddsi_discovery.
