+++ src/core/ddsi/src/ddsi__discovery.h
@@ -30,6 +30,30 @@ struct ddsi_xevent;
 struct ddsi_xpack;
 struct ddsi_domaingv;
 
+typedef enum ddsi_sedp_kind {
+  SEDP_KIND_READER,
+  SEDP_KIND_WRITER,
+  SEDP_KIND_TOPIC
+} ddsi_sedp_kind_t;
+
+/** @component discovery */
+struct ddsi_writer *ddsi_get_sedp_writer (const struct ddsi_participant *pp, unsigned entityid)
+  ddsrt_nonnull_all;
+
+/** @component discovery */
+struct ddsi_proxy_participant *ddsi_implicitly_create_proxypp (struct ddsi_domaingv *gv, const ddsi_guid_t *ppguid, ddsi_plist_t *datap /* note: potentially modifies datap */, const ddsi_guid_prefix_t *src_guid_prefix, ddsi_vendorid_t vendorid, ddsrt_wctime_t timestamp, ddsi_seqno_t seq)
+  ddsrt_nonnull_all;
+
+/** @component discovery */
+bool ddsi_check_sedp_kind_and_guid (ddsi_sedp_kind_t sedp_kind, const ddsi_guid_t *entity_guid)
+  ddsrt_nonnull_all;
+
+/** @component discovery */
+bool ddsi_handle_sedp_checks (struct ddsi_domaingv * const gv, ddsi_sedp_kind_t sedp_kind, ddsi_guid_t *entity_guid, ddsi_plist_t *datap,
+    const ddsi_guid_prefix_t *src_guid_prefix, ddsi_vendorid_t vendorid, ddsrt_wctime_t timestamp,
+    struct ddsi_proxy_participant **proxypp, ddsi_guid_t *ppguid)
+  ddsrt_nonnull_all;
+
 /** @component discovery */
 int ddsi_builtins_dqueue_handler (const struct ddsi_rsample_info *sampleinfo, const struct ddsi_rdata *fragchain, const ddsi_guid_t *rdguid, void *qarg);
